+++ src/common/localStorage.ts
@@ -1,7 +1,6 @@
 export enum LocalStorageKeys {
   FilterState = 'AC_FILTER_STATE',
   OpenWorkItem = 'AC_OPEN_WORK_ITEM_DIALOG',
-  UndoCompleted = 'AC_UNDO_COMPLETED',
   ShowCompletedWi = 'AC_SHOW_COMPLETED_WIS',
   NewStateFlow = 'AC_NEW_FLOW'
 }